Lift setup() helper to standlaone file
